ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
scikit-learn==1.5.2
pytesseract==0.3.13
Pillow==11.0.0
python-multipart==0.0.9

